Refactor and extend data model
This commit adds a new entity class, IlluminaSeqPlatformUnit, that serves as an intermediate layer between the IlluminaSeqRun/SeqLibrary and IlluminaSeqReads entity classes, taking over some of the functionality previously in IlluminaSeqReads.  It also adds a new entity class, IlluminaSeqReadsSet, that refers to a collection of IlluminaSeqReads objects (e.g., a batch of FASTQ files).

The following changes were made to the data model:
* IlluminaSeqPlatformUnit, rather than IlluminaSeqReads, now refers to IlluminaSeqRun and SeqLibrary entities.
* IlluminaSeqPlatformUnit, rather than IlluminaSeqReads, now contains the information about the sequencing lane(s) used to generate the reads.
* The 'flowcell' and 'barcodes' fields have been copied from the IlluminaSeqRun and SeqLibrary entity classes, respectively, into the IlluminaSeqPlatformUnit entity class.
* IlluminaSeqReads now refers only to IlluminaSeqPlatformUnit entities and to FASTQ and/or uBAM WorkFiles.
* IlluminaSeqReadsSet refers to one or more IlluminaSeqReads objects and associates it with a descriptive name.